Request user password to edit user info
This is needed because with authelia we don't have any more the user password after the login and so we don't have the password to login in LDAP to update the user info. So we need to request the password on user update to have a way to authenticate to LDAP. On other way this bring more security when the user update his informations.